[AIRFLOW-YYYY] Rename core.py to test_core.py
