+++ scripts/ci/pylint_todo.txt
@@ -321,7 +321,7 @@
 ./tests/contrib/sensors/test_redis_pub_sub_sensor.py
 ./tests/contrib/sensors/test_sqs_sensor.py
 ./tests/contrib/sensors/test_weekday_sensor.py
-./tests/core.py
+./tests/test_core.py
 ./tests/executors/test_base_executor.py
 ./tests/executors/test_celery_executor.py
 ./tests/executors/test_dask_executor.py
