+++ tests/test_utils/mocks.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -17,7 +15,6 @@
 # specific language governing permissions and limitations
 # under the License.
 
-# flake8: noqa
+from unittest import mock
 
-from .api import *  # type: ignore
-from .test_core import *  # type: ignore
+send_email_test = mock.Mock()
